+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * #291 Add configurable closing options for Modal (thanks @jwahdatehagh)
 * #298 Fix mobile detection datepicker for SSR
 * #302 Fix datepicker calendar display when disabled
+* #305 Fix datepicker calendar display when disabled and inline
 
 ## 0.5.1